forgot pass done

+++ app/Http/Controllers/Api/ForgetPasswordController.php
@@ -60,12 +60,20 @@ class ForgetPasswordController extends Controller
         if (!$updatePassword) {
             return back()->withInput()->with('error', 'Invalid token!');
         }
-
-        $user = User::where('email', $request->email)
+        $is_email_exist = User::where('email', $request->email)->exists();
+        if($is_email_exist){
+            $user = User::where('email', $request->email)
             ->update(['password' => Hash::make($request->password)]);
 
-        DB::table('password_resets')->where(['email' => $request->email])->delete();
+            DB::table('password_resets')->where(['email' => $request->email])->delete();
 
-        return view('auth.redirectToLogin');
+            return view('auth.redirectToLogin');
+        }else{
+            return response()->json([
+                'message'=>'Email is not valid',
+                'status'=>500,
+            ]);
+        }
+        
     }
 }
